+++ src/python/architectures/__init__.py
@@ -1,0 +1,1 @@
+from .image_architectures import get_im_clf_model, get_im_sgm_model, get_architectures_by_type
