+++ src/python/datasets/__init__.py
@@ -1,0 +1,1 @@
+from .image_datasets import ImageClassificationDataset, ImageSegmentationDataset
