+++ src/python/models/__init__.py
@@ -1,0 +1,1 @@
+from .image_models import ImageClassificationModel, ImageSegmentationModel
