+++ src/python/trainers/__init__.py
@@ -1,0 +1,1 @@
+from .image_trainers import ImageClassificationTrainer, ImageSegmentationTrainer
